outsourced text_input, ask_bool and ask_text
